+++ legacy/account/src/main/java/app/k9mail/legacy/account/Account.kt
@@ -600,12 +600,6 @@ open class Account(
         return uuid.hashCode()
     }
 
-    enum class ShowPictures {
-        NEVER,
-        ALWAYS,
-        ONLY_FROM_CONTACTS,
-    }
-
     enum class QuoteStyle {
         PREFIX,
         HEADER,
