+++ legacy/account/src/main/java/app/k9mail/legacy/account/LegacyAccountWrapper.kt
@@ -3,7 +3,6 @@ package app.k9mail.legacy.account
 import app.k9mail.legacy.account.Account.Companion.NO_OPENPGP_KEY
 import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
-import app.k9mail.legacy.account.Account.ShowPictures
 import app.k9mail.legacy.notification.NotificationSettings
 import com.fsck.k9.mail.ServerSettings
