+++ legacy/account/src/main/java/app/k9mail/legacy/account/ShowPictures.kt
@@ -1,0 +1,7 @@
+package app.k9mail.legacy.account
+
+enum class ShowPictures {
+    NEVER,
+    ALWAYS,
+    ONLY_FROM_CONTACTS,
+}
